add metadata for geo works and put state behavior in separate module
