+++ app/models/image_work.rb
@@ -3,6 +3,8 @@ class ImageWork < ActiveFedora::Base
   include ::GeoConcerns::ImageWorkBehavior
   include ::CurationConcerns::BasicMetadata
   include ::GeoConcerns::BasicGeoMetadata
+  include ::GeoMetadata
+  include ::StateBehavior
   self.valid_child_concerns = [RasterWork]
 
   # Use local indexer
