+++ app/models/raster_work.rb
@@ -4,6 +4,8 @@ class RasterWork < ActiveFedora::Base
   include ::CurationConcerns::BasicMetadata
   include ::GeoConcerns::BasicGeoMetadata
   include ::GeoConcerns::GeoreferencedBehavior
+  include ::GeoMetadata
+  include ::StateBehavior
   self.valid_child_concerns = [VectorWork]
 
   # Use local indexer
