+++ app/models/scanned_resource.rb
@@ -4,6 +4,7 @@ class ScannedResource < ActiveFedora::Base
   include ::CurationConcerns::WorkBehavior
   include ::CurationConcerns::BasicMetadata
   include ::CommonMetadata
+  include ::StateBehavior
   include ::StructuralMetadata
   include ::HasPendingUploads
   include ::CollectionIndexing
