+++ spec/factories/vector_works.rb
@@ -1,6 +1,7 @@
 FactoryGirl.define do
   factory :vector_work do
     title ["Test title"]
+    rights_statement "http://rightsstatements.org/vocab/NKC/1.0/"
     visibility Hydra::AccessControls::AccessRight::VISIBILITY_TEXT_VALUE_PUBLIC
 
     transient do
